requirements.txt
torch==2.2.2
transformers==4.20.1
Pillow==9.1.1
tqdm==4.64.0
json5==0.9.5
scikit-image==0.19.3
nltk==3.7
gensim==4.0.1
pytube==12.1.0
